image id is treated as u32 everywhere but was actually usize somehow

+++ src/MultiAtlas.zig
@@ -22,7 +22,7 @@ test {
     std.testing.refAllDecls(@This());
 }
 
-pub const ImageId = usize;
+pub const ImageId = u32;
 
 pub const ImageLocation = struct {
     atlas_index: usize,
